+++ src/app/components/modals/atm-task-execution-pods-activity-modal/events-table/header-row.js
@@ -1,0 +1,21 @@
+/**
+ * Shows events table header row.
+ *
+ * @module components/modals/atm-task-execution-pods-activity-modal/events-table/header-row
+ * @author Michał Borzęcki
+ * @copyright (C) 2021 ACK CYFRONET AGH
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
+import Component from '@ember/component';
+import I18n from 'onedata-gui-common/mixins/components/i18n';
+
+export default Component.extend(I18n, {
+  tagName: 'tr',
+  classNames: ['events-table-header-row'],
+
+  /**
+   * @override
+   */
+  i18nPrefix: 'components.modals.atmTaskExecutionPodsActivityModal.eventsTable.headerRow',
+});