+++ src/app/locales/en/components/modals.js
@@ -1,5 +1,7 @@
+import atmTaskExecutionPodsActivityModal from './modals/atm-task-execution-pods-activity-modal';
 import runWorkflowModal from './modals/run-workflow-modal';
 
 export default {
+  atmTaskExecutionPodsActivityModal,
   runWorkflowModal,
 };