+++ src/app/locales/en/components/modals/atm-task-execution-pods-activity-modal/events-table.js
@@ -1,0 +1,6 @@
+import headerRow from './events-table/header-row';
+
+export default {
+  headerRow,
+  noEvents: 'No events to show.',
+};