+++ src/app/locales/en/components/modals/atm-task-execution-pods-activity-modal/pods-table/header-row.js
@@ -1,0 +1,12 @@
+export default {
+  columnLabel: {
+    name: 'Pod name',
+    readiness: 'Readiness',
+    status: 'Status',
+    statusSince: 'Status since',
+  },
+  filter: {
+    current: 'Current',
+    all: 'All',
+  },
+};