+++ src/app/services/mocks/onedata-graph.js
@@ -25,6 +25,7 @@ import { entityType as transferEntityType } from 'oneprovider-gui/models/transfe
 import { entityType as datasetEntityType } from 'oneprovider-gui/models/dataset';
 import { entityType as fileEntityType } from 'oneprovider-gui/models/file';
 import { entityType as archiveEntityType } from 'oneprovider-gui/models/archive';
+import { entityType as atmTaskExecutionEntityType } from 'oneprovider-gui/models/atm-task-execution';
 
 const messagePosixError = (errno) => ({
   success: false,
@@ -621,7 +622,7 @@ const atmStoreHandlers = {
     const endPositon = Math.min(startPosition + limit, maxEntriesCount);
     const storeEntries = [];
     const entryGeneratorFunc = entityId.startsWith('auditLog') ?
-      generateAuditLogContentEntry : generateStoreContentEntry;
+      generateStoreAuditLogContentEntry : generateStoreContentEntry;
 
     for (let i = startPosition; i < endPositon; i++) {
       storeEntries.push(entryGeneratorFunc({ startPosition, index: i }));
@@ -631,6 +632,15 @@ const atmStoreHandlers = {
       list: storeEntries,
       isLast: startPosition >= maxEntriesCount,
     };
+  },
+};
+
+const atmTaskExecutionHandlers = {
+  openfaas_function_pod_event_log(operation, entityId, data, authHint, aspectId) {
+    if (operation !== 'get' || !aspectId) {
+      return messageNotSupported;
+    }
+    return generateJsonInfiniteLogEntries('openfaasActivity', data, { podId: aspectId });
   },
 };
 
@@ -823,6 +833,7 @@ export default OnedataGraphMock.extend({
       // Using entity type string directly, because op_atm_store does not have
       // dedicated model in ember data.
       op_atm_store: atmStoreHandlers,
+      [atmTaskExecutionEntityType]: atmTaskExecutionHandlers,
     });
     this.set(
       'handlers',
@@ -1054,7 +1065,7 @@ function atmWorkflowExecutionSummaryToAttrsData(record) {
   });
 }
 
-function generateAuditLogContentEntry({ index }) {
+function generateStoreAuditLogContentEntry({ index }) {
   return {
     index: String(index),
     success: index % 20 !== 19,
@@ -1084,4 +1095,57 @@ function generateStoreContentEntry({ startPosition, index }) {
     value: value,
     error: { id: 'forbidden' },
   };
+}
+
+function generateJsonInfiniteLogEntries(type, pagingParams, extraData = {}) {
+  const firstEntryDate = new Date();
+  firstEntryDate.setMinutes(0, 0, 0);
+  const firstEntryTimestamp = Math.floor(firstEntryDate.valueOf() / 1000) - 3600;
+  const {
+    index,
+    offset,
+    limit,
+  } = pagingParams;
+  const startPosition = Math.max((Number(index) || 0) + offset, 0);
+  const startTimestampOffset = startPosition * 10;
+  const startEntryTimestamp = firstEntryTimestamp + startTimestampOffset;
+  const nowTimestamp = Math.floor(new Date().valueOf() / 1000);
+  const entries = [];
+  let entryIdx = startPosition;
+  for (
+    let i = startEntryTimestamp; i <= nowTimestamp && entries.length < limit; i += 10
+  ) {
+    entries.push(generateJsonInfiniteLogEntry({
+      index: String(entryIdx),
+      timestamp: i,
+      type,
+      extraData,
+    }));
+    entryIdx++;
+  }
+
+  return {
+    logEntries: entries,
+    isLast: entries.length < limit,
+  };
+}
+
+function generateJsonInfiniteLogEntry({ index, timestamp, type, extraData }) {
+  const entry = {
+    index,
+    // infinite log timestamps are in milliseconds
+    timestamp: timestamp * 1000,
+  };
+  if (type === 'openfaasActivity') {
+    const podId = extraData && extraData.podId;
+    if (!podId) {
+      throw new Error('Pod ID not specified!');
+    }
+    entry.content = {
+      type: 'Normal',
+      reason: 'Completed',
+      message: `Message of event at ${timestamp} (${podId}) (index: #${index})`,
+    };
+  }
+  return entry;
 }